Issue/11 page creation (#27)
* Adding images

Adding Deftech and NSIN logos to the public folder

* Adding sponsor logos

made a footer to add sponsor logos

* changed default font to black

* Created Landing page

1. Created the main body of text for our page.
2. Made a footer that includes the logo for both NSIN and DEFTECH that
2a. links to the respective websites
2b. can adjust and resize themselves according to page size

* creating education page

setting up education page

* setting up page routing paths

creating pages for each education sub section so i can later make the education bullet points in the education page link to each of these pages

* set up initial navbar links

set up a const link and pathname so that the navbar is simple for Christian to set up

* changed navbar to a tsx

changed from .ts to .tsx, shouldnt change anything but not having auto complete was annoying

* page links

Setting up page links so each bullet point links to the correct page

* completing CFR 15 page

Adding initial information to our CFR 15 Page

* our logo link

made our logo link back to the home page

* Changed logo

Old logo had a white background so I swapped it for one that had a transparent background which is better for our background colors

* changed main background color

changed from white to dark grey

* changed sponsor text color

made the "Sponsors:" text the white default color

* changed color scheme

Made the color scheme reflect the figma with the dark grey background and dark blue header with a whit drop shadow to separate the two

* changed text and fixed footer

- removed excess text and simplified what was left
- Fixed footer to bottom of page

* Fixed logo image link +comments

The image logo kept expanding past the logo to the entire line that it was on. by adding className="inline-block" to the <link> component i fixed it

also added comments for readability

* notes for navbar+message for christian

added notes to navbar and a message for christian

* moved up text

moved up text by 1 line

* increased the drop shadow size

* Create searchbar.tsx

* adding searchbar to header layout

making searchbar permenant for the layout on all pages

* creating search page

made search page and added search bar to it

* added search bar icon

put in public folder

* added searchbar to home

working searchbar is now on the home page

* changed routing and conditions

- made it so searching always routes to the search page first before displaying results no matter what path the search was made from

- changed the search function to only search once enter was clicked instead of off of any change

* added search handler

needed for client side search states. search states are default, success, and failure as described in figma.

didnt want to "use client" in a front facing page for security reasons so i made it a separate component, also made it easier and more organized this way

* imported component and added comments

descriptive comments added

imported "SearchResultsHandler" for client side state management

* text layout

centered and moved text. will remake later towards figma design

* enlarged text

* changed default properties

needed to change this so content would fit 100% of the viewport and keep footers on bottom while rest of the content is fine

* searchbar functionality

now makes a dropdown with the results

* spacing for home page

content should now be properly spaced for all viewports

* changed footer color

had it set to gray for visual testing purposes, changed it to be the same as the body color now

* adding main tags

adding for semantic and spacing reasons

* changing up comments

* made initial reports page