added function to verify companies
checks if a company is in the company table in our database, and if so it will add it checks if the isVerified status is false and if so it will "update" it and set it to true.
If the company isn't in the table at all then it will add it in and then set the isVerified status to true and sets riskscore to 0.