105-reviewer-can-control-a-report (#107)
* added toastify

added toastify for alert popups

* Create VerifyButton.tsx

Created a button that only pops up if the user is a reviewer and will verify a company if clicked

* added verifyUser to data.ts

This will verify the users login state, before this was just done in the dashboard page but since I also added a check in the company page I thought that it might be better to just make this a function in data.ts so we aren't repeating code everywhere and so files aren't getting unnecessarily long because of cookie checks

* added toastify notification to layout

* added function to verify companies

checks if a company is in the company table in our database, and if so it will add it checks if the isVerified status is false and if so it will "update" it and set it to true.
If the company isn't in the table at all then it will add it in and then set the isVerified status to true and sets riskscore to 0.

* replaced the cookie check with a verifyUser call

since i moved the cookie check to data.ts, i removed the repeated code from the company page to shorten the file and make it more readable.

* added a reviewer check and verifycompany button

page now checks for reviewer status first and then if the user is a reviewer adds the verifyButton

* Formatting

* formatting

* cleaned up unneeded code...for formatting